backend/requirements.txt
fastapi
uvicorn
python-dotenv
redis
pymilvus
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
motor
pydantic[email]